server/requirements.txt
django-cors-headers==4.3.0
django-filter==23.3
django-storages==1.14.2
djangorestframework==3.14.0
